make attack parameters modifiable

+++ Assets/Scripts/Survivors/Units/Model/IAttackModel.cs
@@ -4,6 +4,6 @@ namespace Survivors.Units.Model
     {
         float TargetSearchRadius { get; }
         float AttackDistance { get; }
-        int AttackDamage { get; } 
+        float AttackDamage { get; } 
     }
 }
